flask_app/requirements.txt
keras
tensorflow
pillow
pandas
matplotlib
numpy
requests
nltk
plotly
Flask
pyarrow
sklearn
seaborn
spacy
opencv-contrib-python
gensim
pyLDAvis
scipy
regex
python-dotenv